+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixes
 
+* **edit:** stop routing the write-surface prefix note to undo_last_edit
 * **edit:** drop shape-based refusal from content
 
 ### Documentation